Se agregan mas tests de Selenium
