+++ backend/freelancenowbackend/integrationTests/test_navigate_to_card_project.py
@@ -3,7 +3,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 import unittest
-
+#LISTO 1 test
 class ScrollAndClickAutomationCardTest(unittest.TestCase):
 
     def setUp(self):
