+++ backend/freelancenowbackend/integrationTests/test_sort_areas_byname.py
@@ -3,7 +3,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 import unittest
-#LISTO
+#LISTO 1 test
 class SortAreasByNameTest(unittest.TestCase):
 
     USERNAME = "raul.quigua@example.com"
